Add to code sample, tweak spacing.
In the code snippet for using the router inside the `server.js` file, there was no context for where the code would be used.  I wrote in the full code for the file, but this could be changed so long as some context is given in the preceding text.
For me it is easier to show it in code.